+++ R/s3.R
@@ -138,3 +138,61 @@ print.box <- function(x, ...) {
   cat_line("<box>")
   print(unbox(x))
 }
+
+
+#' Create zap objects
+#'
+#' @description
+#'
+#' `zap()` creates a sentinel object that indicates that an object
+#' should be removed. For instance, named zaps instruct [env_bind()]
+#' and [call_modify()] to remove those objects from the environment or
+#' the call.
+#'
+#' The advantage of zap objects is that they unambiguously signal the
+#' intent of removing an object. Sentinels like `NULL` or
+#' [missing_arg()] are ambiguous because they represent valid R
+#' objects.
+#'
+#' @param n Either the length of the zap list, or a character vector
+#'   of names for the zaps. In the latter case, the length of the list
+#'   is deduced from the length of the character vector.
+#' @param x An object to test.
+#'
+#' @export
+#' @examples
+#' # Create one zap object:
+#' zap()
+#'
+#' # Create a list of zaps:
+#' zaps(2)
+#' zaps(letters[1:3])
+zap <- function() {
+  `zap!`
+}
+#' @rdname zap
+#' @export
+zaps <- function(n) {
+  if (is_integerish(n, 1L)) {
+    return(rep_len(list(`zap!`), n))
+  }
+
+  if (is_character(n)) {
+    zaps <- rep_len(list(`zap!`), length(n))
+    return(set_names(zaps, n))
+  }
+
+  abort("`n` must be a length or a character vector of names")
+}
+#' @rdname zap
+#' @export
+is_zap <- function(x) {
+  inherits(x, "rlang_zap")
+}
+
+`zap!` <- structure(list(), class = "rlang_zap")
+
+#' @export
+print.rlang_zap <- function(x, ...) {
+  cat_line("<zap>")
+}