+++ tests/testthat/test-s3.R
@@ -73,3 +73,16 @@ test_that("as_box_if() ensures boxed value if predicate returns TRUE", {
 test_that("unboxing a non-boxed value is an error", {
   expect_error(unbox(NULL), "must be a box")
 })
+
+test_that("zap() creates a zap", {
+  expect_is(zap(), "rlang_zap")
+  expect_true(is_zap(zap()))
+})
+
+test_that("zaps() create many zaps", {
+  expect_identical(zaps(2), list(zap(), zap()))
+  expect_identical(zaps(c("foo", "bar")), list(foo = zap(), bar = zap()))
+  expect_true(every(zaps(2), is_zap))
+  expect_error(zaps(1:2), "must be a length or a character vector of names")
+  expect_error(zaps(env()), "must be a length or a character vector of names")
+})